op-e2e: add test option to utilize erigon for L2
This change is an attempt to include op-erigon in the op-e2e tests in a
minimally invasive way.  Almost all of the tests are unchanged, with
some small structural changes to the setup to allow for utilization of a
Geth or Erigon client as the L2 Ethereum client.

The op-erigon directory is new, and is an unfortunate hack to get Erigon
to build inside of the op-e2e tests.  Unfortunately, because Erigon and
Geth share many dependencies (and much of their code), it's seemingly
impossible to build a single Go binary containing both op-geth and
op-erigon.  Consequently this change takes the approach of building an
Erigon binary, and running it in a separate process.  The resulting test
path is slower and more resource intensive than the Geth path, but,
utilizing parallelism the test suite still successfully executes in
approximately 120 seconds on this laptop, compared with about 90 seconds
for the Geth tests.

To execute these tests simply run `make test-erigon` in the `op-e2e`
directory.  Or, alternatively pass the `--erigon` flag to your `go test`
command.